requirements.txt
streamlit
pandas
Pillow
numpy
easyocr
pytesseract
